Added some basic rmsd calculation
RMSD, also known as root mean squared deviation, is a metric to describe how
different the discrete values of a vector are with repspect to an equally long
vector. In our case, we can compare vectors of any value from day to day or from
week to week. A cursory look seems to suggest that abnormal days of traffic are
a multi-day occurance, with traffic abnormalities on a Thursday creating traffic
abnormalities on the following Friday. A full rundown of the results are on the
way.