Added RMSD calculations vs first week
I added a section that compares the RMSD of each week against the first week.
This helps cut down on some RMSD echo that you experience in the week-by-week
data. Additionally, there almost seems to be a concerted motion, in so far as
multiple days of a week experienceing a similar change in RMSD. This may
suggest that each day's traffic is connected.