Fixed labeling errors in RMSD documentation
There some missing labels and spelling mistakes in the RMSD report. I fixed the
spelling errors and added labels to figures where necessary. Additionally, I
added some plots for RMSD calculations in comparision to the first week in
addition to the current plots that calculate RMSD with the prior week.