Added pairwise RMSD calculations
I've put together a pairwise calculations of RMSD values for each day. Some
preliminary analysis suggests that this is a good way of picking up specific
days that are odd. It seems to detect holidays and weekends as being abnormal.
We may need to look at intermediate deviations to detect true abnormalities.