Added Correlation pattern analyses.
I added a script that performs an analysis similar to the RMSD analysis but
using the Pearson Correlation coffeicient. This seems to be as robust as
RMSD but it has the benifit of being widely understood. The analysis picks
up holidays and the odd sunday traffic very easily. I think that it may be
useful to point out non-holiday abnormalities for each day of the week with the
intention of tracking down a causative factor.

What do you folks think?