Added a proper title page and some scafolding
I added a nice looking title page to the document and added some basic outline
to the sections. A good reference for an existing traffic analysis can be found
here: http://www.muni.org/Departments/works/traffic/Documents/C_and_Klatt_Traffic_Analysis_Report.pdf
It is real analysis, so there isn't a lot of statistics or mathematics in it,
but I think some of the general content may be useful.