Addition of basic trend analysis to report
I added some basic trend analysis to the report. The major findings is that
there is a dip in traffic during the winter months, but an overall increase in
the traffic volume and the delay at the intersection over time. Right now there
is a lot of lorem lpsum filler text, but it is a start.