Fleshed out the report more. Added some figures.
I've added a lot more scafolding text to fill in the sections, as well as
starting to add in the content and figures for the traffic analysis section.

I've implemented a format that may be benificial, in which in the introduction
we discuss the intersection itself, with satellite imagery and depicitions
of the intersection conditions. We should also outline the goals of the report
in the introduction.

We then continue into the data analysis portion of the
report. We begin by outlining the traffic-related analysis of the intersection,
includeing the intersection Level of Service (LOS), the traffic volume and
density, and the traffic velocity. This section should include the Fundamental
Diagram, which compares traffic flow (volume) with traffic density. With this
diagram, we can determine the free-flow velocity and a typical traffic wave
velocity for the intersection. These aren't particularly impressive
mathematically, but they are actually relevent to the DOT.

Next, we discuss trends and patterns. This can include changes in traffic
volume, stops, delay, etc. over time. Additionally, this section will outline
abnormalities in the traffic includeing holidays and other phenomina. We can
use this as an opportunity to show examples of interesting patterns of traffic
utilization. I've also had minor luck with Fast Fouier Transformations, which
provide some non-trvial periodicities in the traffic (odd periods in addition
to daily and weekly periods). We could also explore PCA, which suggests that
there is a principle component that is highly responsible for traffic patterns.
This may be time, but my experience with DateTime objects and PCA is extremely
limited.

We can then move into other analyses, such as Bayesian prediction of congestion
based on time and day as well as weather-based analyses. Recommendations can
then be some sort of useful or inciteful remark about how there is a lot of
traffic.